backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
pillow==10.1.0
requests==2.31.0
pydantic==2.5.0
